requirements.txt
asyncio==3.4.3
attrs==20.3.0
autopep8==2.0.2
awscli==1.28.0
backports.shutil-get-terminal-size==1.0.0
backports.shutil_which==3.5.2
beautifulsoup4==4.12.2
black==23.3.0
blinker==1.6.2
botocore==1.30.0
braceexpand==0.1.7
Brotli==1.0.9
cachelib==0.10.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
check50==3.3.7
cli50==7.4.0
click==8.1.3
colorama==0.4.4
compare50==1.2.6
contourpy==1.1.0
cryptography==41.0.1
cs50==9.2.5
CS50-VSIX-Client @ file:///opt/cs50/extensions/cs50vsix-client
cssselect2==0.7.0
cycler==0.11.0
defusedxml==0.7.1
docutils==0.16
EditorConfig==0.12.3
Flask==2.3.2
Flask-Session==0.5.0
fonttools==4.40.0
fpdf2==2.7.4
greenlet==2.0.2
help50==3.0.5
html5lib==1.1
icdiff==2.0.6
idna==3.4
inflect==7.0.0
iniconfig==2.0.0
intervaltree==3.1.0
itsdangerous==2.1.2
jellyfish==0.11.2
Jinja2==3.1.2
jmespath==1.0.1
jsbeautifier==1.14.8
kiwisolver==1.4.4
lib50==3.0.6
MarkupSafe==2.1.3
matplotlib==3.7.2
mypy==1.4.1
mypy-extensions==1.0.0
natsort==8.4.0
numpy==1.25.0
packaging==23.1
pathspec==0.11.1
pexpect==4.8.0
Pillow==10.0.0
platformdirs==3.8.1
pluggy==1.2.0
ptyprocess==0.7.0
pyasn1==0.5.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.11
pydyf==0.7.0
Pygments==2.15.1
pyparsing==3.0.9
PyPDF2==2.12.1
pyphen==0.14.0
pytest==7.4.0
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2023.3
PyYAML==5.4.1
render50==9.1.2
requests==2.31.0
rsa==4.7.2
s3cmd==2.3.0
s3transfer==0.6.1
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==1.4.46
sqlparse==0.4.4
style50==2.7.5
submit50==3.1.1
termcolor==1.1.0
tinycss2==1.2.1
tqdm==4.65.0
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==1.26.16
weasyprint==59.0
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.3.6
zopfli==0.2.2
